Handling client disconnections and adding static filter for sorted fuzzer
This commit includes fixes for better handling disconnected clients, by trying to reconnect them.

For the sorted fuzzer, a static filter for methods containing "close" was implemented, because these methods likely close the context handle and cause the rest of the procedure calls to fail that take the context handle as input parameter.

Finally, a bug was fixed that the endpoint for ncacn_tcp was not included when a interface supports both ncalrpc and ncacn_tcp, GetRpcData.